RSC-8 refactor(project): edit small mistakes and move functional to the right folders and files
