+++ src/types.ts
@@ -1,3 +1,5 @@
+import { IAccount, ICategory, ITransaction } from './interfaces';
+
 export type IChart = {
   labels: string[];
   datasets: {
@@ -6,3 +8,7 @@ export type IChart = {
     backgroundColor: string[];
   }[];
 };
+
+export type Anchor = 'top' | 'left' | 'bottom' | 'right';
+
+export type DataAllFB = Partial<IAccount> | Partial<ICategory> | Partial<ITransaction>;
